requirements.txt
-e git+https://github.com/Zawala/aidflow.git@8d4a11eb19999ea1577c347428fb3002183c94c1#egg=aidflow
annotated-types==0.7.0
asttokens==2.4.1
attrs==25.1.0
Babel==2.13.1
backcall==0.2.0
backoff==2.2.1
barcodenumber==0.5.0
beautifulsoup4==4.12.3
bleach==6.0.0
bleach-allowlist==1.0.3
blis==1.2.0
boto3==1.34.162
botocore==1.34.162
braintree==4.20.0
Brotli==1.1.0
cachetools==4.2.4
cairocffi==1.5.1
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
chardet==5.1.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.21.0
confection==0.1.5
coverage==6.5.0
croniter==2.0.7
cryptography==44.0.2
-e git+https://github.com/Zawala/cso_database.git@a0bac908087f827463670b5bc07be9892f502277#egg=cso_database
cssselect==1.2.0
cssselect2==0.7.0
cssutils==2.9.0
cymem==2.0.11
dataclasses-json==0.5.14
decorator==5.1.1
docopt==0.6.2
dropbox==11.36.2
email-reply-parser==0.5.12
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
-e git+ssh://git@bitbucket.org/logikmind/erp_overrides.git@01b5af9b2c9eee125d2d76a6a2c5867befdffef9#egg=erp_overrides
-e git+https://github.com/frappe/erpnext.git@2ce8299bc8c5dbcf99dea5997ae9000ab7645a6d#egg=erpnext
et-xmlfile==1.1.0
executing==2.1.0
Faker==18.10.1
filelock==3.13.4
filetype==1.2.0
fonttools==4.54.1
-e git+https://github.com/frappe/frappe.git@7d580cf4f189858b51f9108bced8de2a724cff8a#egg=frappe
frappe-bench==5.23.0
-e git+https://github.com/shridarpatil/frappe_whatsapp@66599ca09ed081d4337c890b32b2087d89aba4db#egg=frappe_whatsapp
freezegun==1.2.2
fsspec==2025.3.0
gitdb==4.0.11
GitPython==3.1.43
gocardless_pro==1.22.0
google-api-core==1.34.1
google-api-python-client==2.2.0
google-auth==1.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.65.0
googlemaps==4.10.0
gunicorn @ git+https://github.com/frappe/gunicorn@72c1e495d89259151e73947a057432d528b06bb0
hiredis==2.2.3
holidays==0.57
honcho==2.0.0
-e git+https://github.com/frappe/hrms.git@461897bf67966657b3eb0491113713d38f6c1e07#egg=hrms
html5lib==1.1
httplib2==0.22.0
huggingface-hub==0.29.3
hypothesis==6.77.0
idna==3.10
ipython==8.15.0
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
ldap3==2.9.1
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
markdown2==2.4.13
markdownify==0.11.6
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib-inline==0.1.7
maxminddb==2.6.2
maxminddb-geolite2==2018.703
mdurl==0.1.2
monotonic==1.6
murmurhash==1.0.12
mypy-extensions==1.0.0
nltk==3.9.1
num2words==0.5.13
numpy==2.2.4
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
-e git+https://github.com/frappe/payments.git@fb7c2a93cce16f19d41dfc9bdfd93a779c2e6a9a#egg=payments
paytmchecksum==1.7.0
pdfkit==1.0.0
pexpect==4.9.0
phonenumbers==8.13.55
pickleshare==0.7.5
pillow==10.3.0
plaid-python==7.2.1
ply==3.11
posthog==3.0.2
premailer==3.10.0
preshed==3.0.9
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==22.3.5
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
pydyf==0.10.0
Pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.1.1
pyngrok==6.0.0
pyOpenSSL==25.0.0
pyotp==2.8.0
pyparsing==3.1.4
pypdf==3.17.4
pyphen==0.16.0
PyPika==0.48.9
pypng==0.20220715.0
PyQRCode==1.2.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-magic==0.4.27
python-stdnum==1.20
python-youtube==0.8.3
pytz==2023.3
PyYAML==6.0.2
qrcode==8.0
rapidfuzz==2.15.2
rauth==0.7.3
razorpay==1.2.0
redis==4.5.5
regex==2024.11.6
requests==2.32.3
requests-oauthlib==1.3.1
responses==0.23.1
RestrictedPython==8.0
rich==13.9.4
rq==1.15.1
rsa==4.9
s3transfer==0.10.2
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
sentry-sdk==1.37.1
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
smmap==5.0.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sql_metadata==2.11.0
sqlparse==0.5.1
srsly==2.5.1
stack-data==0.6.3
stone==3.3.8
stripe==2.56.0
tenacity==8.2.3
terminaltables==3.1.10
textblob==0.19.0
thinc==8.3.4
threadpoolctl==3.6.0
tinycss2==1.1.1
tokenizers==0.21.1
tomli==2.0.1
tqdm==4.67.1
traceback-with-variables==2.0.4
traitlets==5.14.3
transformers==4.50.3
typer==0.15.2
types-PyYAML==6.0.12.20241230
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.2
Unidecode==1.3.8
unittest-xml-reporting==3.2.0
uritemplate==3.0.1
urllib3==2.2.3
vobject==0.9.9
wasabi==1.1.3
watchdog==3.0.0
wcwidth==0.2.13
weasel==0.4.1
weasyprint==59.0
webencodings==0.5.1
Werkzeug==3.0.4
wheel==0.44.0
Whoosh==2.7.4
wrapt==1.17.2
xlrd==2.0.1
-e git+https://github.com/Zawala/zan_website.git@e4c5ff63de4a46ad61d7ff18b990e1935fd43b3d#egg=zan_website
zopfli==0.2.3
zxcvbn==4.4.28
